Move code out of module scope and into core app ready method. Update model based tests that were dependent on weird import behaviour this induced.
