+++ kolibri/core/test/__init__.py
@@ -1,0 +1,3 @@
+from kolibri.core.test.helpers import setup_test_app
+
+setup_test_app("kolibri.core.test.test_app")
